+++ build.rs
@@ -1,4 +1,5 @@
 fn main() -> Result<(), Box<dyn std::error::Error>> {
+    println!("cargo:rerun-if-changed=proto/cluster.proto");
     tonic_build::compile_protos("proto/cluster.proto")?;
     Ok(())
 }